(My second commit)

+++ week-2/week-2project/project_3.rs
@@ -2,7 +2,7 @@ fn main() {
 	// Given data
 	let principal: f64 = 510_000.0; // #510,000
 	let rate: f64 = 5.0; // 5% per annum
-	let time: f64 = 3; // 3 years
+	let time: i32= 3; // 3 years
 
 	// Calculate Amount
 	let amount = principal * (1.0 - (rate / 100.0)).powi(time);
